refactor(sanity): Move date inputs UI Components to ui-components (#7410)
* refactor(sanity): move base components from form/inputs/DateInputs to ui-components

* refactor(sanity): update the imports
