+++ packages/sanity/src/core/form/inputs/DateInputs/utils.ts
@@ -1,4 +1,4 @@
-import {type CalendarLabels} from './base/calendar/types'
+import {type CalendarLabels} from '../../../../ui-components/inputs/DateInputs/calendar/types'
 
 export function isValidDate(date: Date): boolean {
   return date instanceof Date && !isNaN(date.valueOf())
