+++ packages/sanity/src/core/tasks/components/form/fields/DateEditFormField.tsx
@@ -6,9 +6,9 @@ import {type KeyboardEvent, useCallback, useMemo, useRef, useState} from 'react'
 import ReactFocusLock from 'react-focus-lock'
 
 import {Button, Popover} from '../../../../../ui-components'
+import {type CalendarLabels} from '../../../../../ui-components/inputs/DateInputs/calendar/types'
+import {DatePicker} from '../../../../../ui-components/inputs/DateInputs/DatePicker'
 import {type FormPatch, type PatchEvent, set, unset} from '../../../../form'
-import {type CalendarLabels} from '../../../../form/inputs/DateInputs/base/calendar/types'
-import {DatePicker} from '../../../../form/inputs/DateInputs/base/DatePicker'
 import {getCalendarLabels} from '../../../../form/inputs/DateInputs/utils'
 import {useDateTimeFormat} from '../../../../hooks'
 import {useTranslation} from '../../../../i18n'
