+++ packages/sanity/src/ui-components/inputs/DateInputs/calendar/utils.ts
@@ -1,6 +1,6 @@
 import {addDays, eachWeekOfInterval, getWeek, lastDayOfMonth, startOfMonth} from 'date-fns'
 
-import {useCurrentLocale} from '../../../../../i18n/hooks/useLocale'
+import {useCurrentLocale} from '../../../../core/i18n/hooks/useLocale'
 import {TAIL_WEEKDAYS} from './constants'
 
 /**
